+++ auth/engine.go
@@ -1,14 +1,11 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package vm
+package auth
 
-import (
-	"github.com/ava-labs/hypersdk/chain"
-)
+import "github.com/ava-labs/hypersdk/chain"
 
-// TODO: remove orphan chain dependency from VM
-type AuthEngine interface {
+type Engine interface {
 	GetBatchVerifier(cores int, count int) chain.AuthBatchVerifier
 	Cache(auth chain.Auth)
 }
